feat: client-server architecture (#4)
* chore: initial setup

* refactor: create client-server files, enable lib exports

* fix: use appropriate scoping and re-exports

* refactor: move error to error.rs, cleanup lib.rs

* feat: add client.rs for known actions client can take

* feat: serialize actions to store server

* fix: implement action result onto server

When I initially scaffolded this quickly, the actions were on the client which was not quite right. The actions are taken by the client, but they are conducted by the server which maintains the kvstore.

* chore: add log/env_logger tryout

* fix: wip writing to stream for expected outputs

* fix: client expected output for tests

* feat: satisfy KvsEngine trait

* chore: remove env_logger

* chore: re-use tracing instead of env_logger

* feat: add engine check

* refactor: move engine check into KvStore

* refactor: remove file_index, use timestamp

* chore: most compaction logs to debug

These logs were at info level and was leading to quite a verbose output

* docs: update comments

* docs: update readme

* fix: remove constant load of keydir

The loading of the keydir in various places was done to help with the fact that previously the store was held in-memory, so the store needed to be loaded in different places. Now it runs as a server and is a long-running process which is interacted with via a network call, this is no longer required

* feat: add basic benches

* feat: add file handle for keydir and active_log

Before this, I was continually opening to file all the time on every read/write

* fix: kv.get uses keydir

This was always assuming that the entry was in the active file. That is not the case.

* chore: rename removal vars

* refactor: split out compaction for simplicity

Perhaps not the most performant compaction algorithm, but it demonstrates the point

* feat: pass tracing guard through to store from server

* feat: expand errors for no log file or keydir

* chore: remove unnecessary &mut self for log_file name gen

* feat: write path performance improvement

Increase write path performance by not serializing the entire keydir to disk upon every write, instead we only need to serialize it after a set of operations. This has been encompassed to the commit_keydir func and Drop, whereby the state should be stored on the KvStore going out of scope