feat: use write_at for offset position within append_to_log
By taking the positional offset to write at within the appending func, this fixes the concurrent_set function. Previously, we were writing to the same offset multiple times which caused get() to return the incorrect data.

NOTE: There is still a requirement to fix a memory allocation problem.
The program crashes when `cargo test` is ran altogether.