feat: wip for load func
Build the keydir from log files on disk. Prior to this we were
serializing the entire keydir on `Drop` which was a major bottleneck.
Instead we can rebuild the directory by reading from log files.

If we're serializing and loading an entire key directory on disk
then the WAL doesn't entirely serve a purpose since it is unused,
we simply have a k-v store already on disk which we continually
(de)serialize every single time we need too. With this new WIP
implementation, the idea is that we should be able to crash in the
middle of operations and then rebuild the index from log files that
were persisted.