feat: use entry api for index rebuild
Prior to this, we were inserting every single element into the keydir
from scratch which meant that every item in every log file was placed
into the index, we don't need to do this.

Instead, we should only modify the entry if it exists and the timestamp
of the log entry is greater than the one we already have, i.e. it is
more up to date.