feat: introduce tokio (#7)
* chore: project setup

* chore: naive threadpool

* feat: wip use dashmap for concurrent access for reads/write

We cannot have multiple mutable references to the same data for both set/remove functions, this requires a change from BTreeMap to the concurrent DashMap (or crossbeam alternative) as the insert/remove functions here do not take &mut self and therefore do not force the implementation to be mutable either

* feat: wip rewrite compaction to singular file with size limit

* feat: use internal Writer with RefCell

* fix: compaction

The timestamps were being set incorrectly because they were always being generated from now(), which meant that the timestamp for the keydir was always later than the entry into the WAL. During comparison this made it so that the items were never compacted

* refactor: use ? operator for std::io errors

* feat: wip concurrent set/get

This requires some extra thinking. I believe we can remove the RwLock around the structures which are already wrapped in one to avoid extra verbosity/unnecessary locking all over the place.

We need to consider how the file lock is handled across threads. Pushing to have this remotely

* feat: insert into keydir before compaction

* chore: use .log in file name

This was part of the name before, but it makes more sense to be the file extension

* fix: compaction

Compaction was broken because the opened file was never seeked to the offset, it was simply reading key0 every single time in the log file

* feat: maintain active files and handles to remove unnecessary files

* fix: pub Result type

* chore: remove benches to align with project-spec

* feat: use Arc<DashMap> for multi-thread safety

* chore: wip concurrent writes/reads

* feat: use bufwriter for compaction file

* feat: use thread_local static for MAX_LOG_FILE_SIZE

* feat: use write_at for offset position within append_to_log

By taking the positional offset to write at within the appending func, this fixes the concurrent_set function. Previously, we were writing to the same offset multiple times which caused get() to return the incorrect data.

NOTE: There is still a requirement to fix a memory allocation problem.
The program crashes when `cargo test` is ran altogether.

* fix: use compaction buffer flush after all entries

Prior to this, the bufwriter wasn't actually doing anything because flush was called after every single write, which meant that the syscall was also done. Now, the writes go into the buffer first and are then flushed at the end of the iteration

* chore: use bufreader for file_handles during compaction

* feat: use naive threadpool

* chore: use rayon threadpool

I'm removing these because it has already been done in the Rust book so I'm not wanting to rehash over it again. I'll utilise the rayon threadpool for speed and move on to the next task

* feat: add basic benches

* fix: scope threads, no cloning in server

* feat: wip for load func

Build the keydir from log files on disk. Prior to this we were
serializing the entire keydir on `Drop` which was a major bottleneck.
Instead we can rebuild the directory by reading from log files.

If we're serializing and loading an entire key directory on disk
then the WAL doesn't entirely serve a purpose since it is unused,
we simply have a k-v store already on disk which we continually
(de)serialize every single time we need too. With this new WIP
implementation, the idea is that we should be able to crash in the
middle of operations and then rebuild the index from log files that
were persisted.

* refactor: logging to use KVS_LOG

Enables ease of setting logging level for tests and server usage

* refactor: use StoreConfig for new KvStore

* chore: load fn use match

* chore: tracing/log output

* revert: benchmark test

* docs: docstring for log_location

* feat: introduce tokio to kvs-client

* refactor: remove keydir path and handles

We no longer serialise the keydir/index, so there is no need to keep a
dangling reference to the location of this file. It would merely serve
as confusion since it points to something which is actually never used
or loaded from the disk - the keydir is rebuilt via log files on crash.

* feat: wip async server

Work in progress for utilising tokio on the server store. This is
panicking on retrieval of an Action from the stream.

Committing here to push remotely.

* wip: tokio server

Continue to modify this for sending/reading the size of the data
before it is serialized, that way we can use read_exact etc.

* feat: use write-size hinting on client/server.

By writing a u64 size hint for the incoming data, we can now use
`read_exact` for the serialisation and fully implement an async
server.

* feat: add serialize_and_hint macro

* refactor: simplify client from macro

* feat: use async tests

Commented the concurrent tests temporarily

* refactor: consolidate new log file functionality

* refactor: convert tests to tokio-compatible

* refactor: use chrono for timestamp gen

* doc: update docstrings for fns

* docs: more info

* feat: convert benches to async

* feat: add debug lines for removal

* feat: add randomised retrieval test

* feat: use entry api for index rebuild

Prior to this, we were inserting every single element into the keydir
from scratch which meant that every item in every log file was placed
into the index, we don't need to do this.

Instead, we should only modify the entry if it exists and the timestamp
of the log entry is greater than the one we already have, i.e. it is
more up to date.

* chore: clippy

* chore: add warn log for load

* ci: add basic workflow

* project: rename to sqrl

* chore: fmt

* docs: update readme

* chore: update LOG_PREFIX

* feat: entry-api for set

* docs: add flake test mention