fix: flakey overwrite_value test
The flakiness was caused because the next timestamp of the log file was
created using microsecond precision. In the test case, this meant that
_sometimes_ it was ran fast enough that the same file which was previous
the active log upon `drop` was then used as the same file again upon
another call to creation a new active log file. This meant that the old
log file was being written to and when we read from the offset of the
value, we got an old value back because it simply hadn't been
overwritten yet.

By changing the log file prefix to use microsecond precision, we fix
this problem because we _actually_ create a new log file on `Drop` in the
test(s) because the timestamp precision is high enough such that the
suffix is different.