refactor: rename ReplicationClient to ReplicatedServer
The initial implementation of this was slightly backwards with the
naming. The ReplicationClient which also served requests does not
make much sense. The better way to understand this is this commit
entails, that is a ReplicatedServer which contains its own store
and a client with an awareness of other server stores.