feat: replicated cache (#16)
* docs: update readme

* refactor: minor rejig in prep for replication crate

* chore: scaffold replication

* docs: update lib

* chore: basic file structure/test

* feat: use client/server wrappers (#17)

* feat: implement wrapper client

* feat: impl Client trait

* feat: generic impl of replication client

* docs: update timestamp docstring for versioning

* wip: replication

* refactor: move KvServer to server.rs

* feat: include timestamp in return value

As part of the replication, we'll need a version/generation number or
similar in order to distinguish between later versions of a value on
various remote stores. We can use the timestamp for this and the
most recent (highest) value will win, i.e. last-write wins.

* feat: include timestamp in return value

As part of the replication, we'll need a version/generation number or
similar in order to distinguish between later versions of a value on
various remote stores. We can use the timestamp for this and the
most recent (highest) value will win, i.e. last-write wins.

* wip: replication

* refactor: rename to StandaloneServer

* refactor: rename ReplicationClient to ReplicatedServer

The initial implementation of this was slightly backwards with the
naming. The ReplicationClient which also served requests does not
make much sense. The better way to understand this is this commit
entails, that is a ReplicatedServer which contains its own store
and a client with an awareness of other server stores.

* wip: replication

* wip: replication

* wip: replication

* wip: replication

Need to fix deadlocks

* wip: replication

Simplify to 3 nodes max for logic and try impl with futures Either/select functionality

* wip: replication

Move to static node count within cluster

* feat: replicate to followers

* fix: use standalone server for tests

The replication with 3 `ReplicatedServer`'s is an infinite loop for
retrieval and setting. The replication strategy should actually be a
single server that is configured to replicate to other nodes, static to
2 for now.

* minor cleanup/rename

* removal for replicated server

* docstring for replication crate

* cargo clippy

* basic test for replication mode cli

* scaffold cli for replication server

* feat: rearchitecture replication to leader/follower

Simplify architecture for replication mode such that we use a simple
leader/follower model. Leader will simply replicate to followers.

This can be altered in the future for smarter replication strategy
or even distribution via consistent hashing etc.